requirements.txt
bidscoin
python-dateutil
parsedatetime
pandas
